free_ssl_service/backend/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
PyMySQL==1.0.2
certbot==1.22.0
python-dotenv==0.19.0
celery==5.2.3
redis==4.2.2
requests==2.26.0
Authlib==0.15.5
PyJWT==2.1.0
sendgrid==6.9.7
Flask-WTF==1.0.0
Flask-Limiter==1.4.0
cryptography==3.4.8
